+++ changelog.md
@@ -1,6 +1,18 @@
 # Simple Site Monitor
 
 
+## v0.1.2 - History Page and UI improvements
+### UI
+- Create "History" page with advanced search and filtering
+- Improve colors on list view tables
+- Modify CSS to use full page width for Monitoring, Sites, Settings and newly created History pages
+- Settings Page improvements including save icon fix and flashing notification when changes are made
+- Move Level of Detail options to Webhooks section
+- Fix table first column alignment
+- Add navigation highlighting for current page
+- Fix automatic tag highlighting on page load
+
+
 ## v0.1.1 - Bug Fixes and Improvements
 ### UI
 - Sites with unknown tag will now appear in monitoring page but will highlight that they the runner hasnt scanned it yet.